move clef to a package
